@@ -1,6 +1,6 @@
 {
 	"name": "svg-chunk-webpack-plugin",
-	"version": "4.0.2",
+	"version": "5.0.0",
 	"description": "Generate SVG sprites according to entrypoint dependencies. Each page only imports its own svgs, wrapped as a sprite and optimized by svgo",
 	"keywords": [
 		"svg",
@@ -24,7 +24,8 @@
 	},
 	"license": "MIT",
 	"author": "Joris DANIEL",
-	"main": "./lib/index.js",
+	"type": "module",
+	"exports": "./lib/index.js",
 	"types": "./types/index.d.ts",
 	"files": [
 		"lib",
@@ -33,7 +34,6 @@
 	"scripts": {
 		"build": "rm -rf ./types ./lib && tsc",
 		"build:example": "rm -rf ./example/dist/ && webpack --config=./example/webpack.config.js --mode=production",
-		"coverage": "jest --coverage && cat ./coverage/lcov.info | coveralls",
 		"dev": "tsc --watch",
 		"dev:unit": "jest --config ./config/jest.config.js --verbose --coverage --watch",
 		"inspect:example": "rm -rf ./example/dist/ && node --inspect node_modules/webpack/bin/webpack.js --config=./example/webpack.config.js --mode=development",
@@ -42,7 +42,7 @@
 		"test": "npm run test:unit",
 		"test:unit": "jest --config ./config/jest.config.js --verbose --coverage"
 	},
-	"prettier": "./config/prettier.config.js",
+	"prettier": "./config/prettier.config.cjs",
 	"dependencies": {
 		"extend": "^3.0.2",
 		"schema-utils": "^4.0.1",
@@ -60,6 +60,7 @@
 		"css-loader": "^6.7.3",
 		"css-minimizer-webpack-plugin": "^5.0.0",
 		"jest": "^29.5.0",
+		"jest-ts-webcompat-resolver": "^1.0.0",
 		"mini-css-extract-plugin": "^2.7.5",
 		"terser-webpack-plugin": "^5.3.7",
 		"ts-jest": "^29.1.0",
@@ -68,7 +69,7 @@
 		"webpack-cli": "^5.0.1"
 	},
 	"peerDependencies": {
-		"webpack": ">= 5.10.3"
+		"webpack": ">= 5.80.0"
 	},
 	"engines": {
 		"node": ">=16.20.0"
