@@ -1,4 +1,4 @@
 import '../svgs/gradient.svg';
 import '../svgs/video.svg';
 
-import './component';
+import './component.js';
